Fix signed/unsigned comparison warning
Changed type of encodedBufferIndex to size_t to eliminate warning. It is
fine as a size_t as its only ever compared to another size_t and used as
an array index.

+++ src/components/rle/RleDecoder.h
@@ -21,7 +21,7 @@ namespace Pinetime {
       const uint8_t* buffer;
       size_t size;
 
-      int encodedBufferIndex = 0;
+      size_t encodedBufferIndex = 0;
       int y = 0;
       uint16_t bp = 0;
       uint16_t foregroundColor = 0xffff;
